Move Tokens into ast module.
