+++ librlox/src/ast/mod.rs
@@ -1,5 +1,6 @@
 pub mod expression;
 pub mod statement;
+pub mod token;
 
 #[cfg(test)]
 mod tests;
