+++ librlox/src/object/mod.rs
@@ -1,0 +1,3 @@
+pub enum Object {
+    Literal,
+}
