+++ librlox/src/scanner/mod.rs
@@ -1,8 +1,6 @@
 pub mod source_scanner;
-pub mod tokens;
 
 pub use source_scanner::{LexResult, Scanner};
-pub use tokens::{Token, TokenType, Value};
 
 #[cfg(test)]
 mod tests;
