+++ librlox/src/scanner/tests/token_lexing/identifiers.rs
@@ -1,4 +1,4 @@
-use crate::scanner::tokens::{TokenType, Value};
+use crate::ast::token::{TokenType, Value};
 
 use super::helpers::{
     compare_single_token_source_helper, compare_single_token_source_with_literal_helper,
